Changed endpoints and added get admin endpoint
